+++ include/aminetxduo/anxnet.h
@@ -24,6 +24,6 @@
 #define ANXNET_CARD_NAMES \
     { "xsurf100", "xsurf", "ariadne2", "hydra", "lanrover", "a2065", \
       "ariadne", "pcmcia", "xsurf500", "3c589", "3ccfem556", "3cxem556", \
-      "genet" }
+      "genet", "zz9000", "zz9000z2" }
 
 #endif /* AMINETXDUO_ANXNET_H */