+++ src/netdev/test/test_netdev_el3.c
@@ -240,6 +240,7 @@ const struct NetdevNicOps netdev_nic_ne2000;
 const struct NetdevNicOps netdev_nic_ed;
 const struct NetdevNicOps netdev_nic_lance;
 const struct NetdevNicOps netdev_nic_genet;
+const struct NetdevNicOps netdev_nic_zz9000;
 
 #include "el3.c"